chore: #29 수정사항 진행
